move pages to component folder
